[setting/applelogin]: 애플 로그인 코드 정리
LoginActivity에 몰려 있는 애플 로그인 코드를 ViewModel로 이동
viewmodel.appleLogin 메서드에 할당
