+++ app/src/main/java/org/gdsc/jmt/view/LoginActivity.kt
@@ -10,15 +10,16 @@ import androidx.activity.result.IntentSenderRequest
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.lifecycle.lifecycleScope
 import org.gdsc.jmt.R
-import org.gdsc.jmt.databinding.ActivityLoginBinding
 import org.gdsc.jmt.viewmodel.LoginViewModel
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.OAuthProvider
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
+import org.gdsc.jmt.databinding.ActivityLoginBinding
 
 class LoginActivity : AppCompatActivity() {
     private lateinit var binding: ActivityLoginBinding
+
     private val loginViewModel: LoginViewModel = LoginViewModel()
     private val TAG = "LoginTest"
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -26,42 +27,11 @@ class LoginActivity : AppCompatActivity() {
         binding = ActivityLoginBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        /// 애플 로그인
+
         binding.appleLoginBtn.setOnClickListener {
-            Log.d(TAG, "appleLoginBtn Clicked")
-            val provider = OAuthProvider.newBuilder(getString(R.string.apple_provider))
-            provider.setScopes(mutableListOf("email", "name"))
-
-            val auth = FirebaseAuth.getInstance()
-
-            val pending = auth.pendingAuthResult
-            if (pending != null) {
-                Log.d(TAG, "peding : not null")
-                pending.addOnSuccessListener { authResult ->
-                    Log.d(TAG, "checkPending:onSuccess:$authResult")
-                    // Get the user profile with authResult.getUser() and
-                    // authResult.getAdditionalUserInfo(), and the ID
-                    // token from Apple with authResult.getCredential().
-
-                    authResult.user?.getIdToken(true)?.addOnSuccessListener {
-                        Log.d(TAG, "checkPending:onSuccess:Token  :  ${it.token}")
-                    }
-                }.addOnFailureListener { e ->
-                    Log.w(TAG, "checkPending:onFailure", e)
-                }
-            }
-
-            // 대기 중 결과가 없다면 실행
-            auth.startActivityForSignInWithProvider(this, provider.build())
-                .addOnSuccessListener { authResult ->
-                    loginViewModel.postAppleToken(authResult.user?.email.toString(), getString(R.string.ios_client_id))
-                }
-                .addOnFailureListener { e ->
-                    Log.w(TAG, "activitySignIn:onFailure", e)
-                }
+            loginViewModel.appleLogin(this)
         }
-
-
+        
         binding.googleLoginBtn.setOnClickListener {
             lifecycleScope.launch(CoroutineExceptionHandler { _, throwable -> throwable.printStackTrace() }) {
                 startForResult.launch(
